+++ GTiff2Tiles.Tests/Tests/EnumsTests.cs
@@ -1,6 +1,7 @@
 ﻿#pragma warning disable CA1031 // Do not catch general exception types
 
 using System;
+using GTiff2Tiles.Core.Enums.Image;
 using NUnit.Framework;
 
 // ReSharper disable UnusedVariable
@@ -14,15 +15,13 @@ namespace GTiff2Tiles.Tests.Tests
         public void Setup() { }
 
         [Test]
-        public void ImageGdal()
+        public void TileExtensions()
         {
             try
             {
-                string lonLat = Core.Constants.Gdal.Gdal.LongLat;
-                string wgs84 = Core.Constants.Gdal.Gdal.Wgs84;
-                string typeByte = Core.Constants.Gdal.Gdal.Byte;
-                string[] repairTifOptions = Core.Constants.Gdal.Gdal.RepairTifOptions;
-                string tempFileName = Core.Constants.Gdal.Gdal.TempFileName;
+                TileExtension png = TileExtension.Png;
+                TileExtension jpg = TileExtension.Jpg;
+                TileExtension webp = TileExtension.Webp;
             }
             catch (Exception exception) { Assert.Fail(exception.Message); }
 
@@ -30,24 +29,11 @@ namespace GTiff2Tiles.Tests.Tests
         }
 
         [Test]
-        public void DateTimePatterns()
+        public void TileType()
         {
             try
             {
-                string longWithMs = Core.Constants.DateTimePatterns.LongWithMs;
-            }
-            catch (Exception exception) { Assert.Fail(exception.Message); }
-
-            Assert.Pass();
-        }
-
-        [Test]
-        public void Extensions()
-        {
-            try
-            {
-                string png = Core.Constants.Extensions.Png;
-                string tif = Core.Constants.Extensions.Tif;
+                TileType raster = Core.Enums.Image.TileType.Raster;
             }
             catch (Exception exception) { Assert.Fail(exception.Message); }